% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{atsu_colors}
\alias{atsu_colors}
\title{Create atsu colors from hex codes}
\format{
An object of class \code{character} of length 37.
}
\usage{
atsu_colors
}
\description{
Create atsu colors from hex codes
}
\keyword{datasets}
