% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{atsu_cols}
\alias{atsu_cols}
\title{Function to extract atsu colors as hex codes}
\usage{
atsu_cols(...)
}
\arguments{
\item{...}{Character names of atsu_colors}
}
\description{
Function to extract atsu colors as hex codes
}
