% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_atsu.R
\name{atsu_pal}
\alias{atsu_pal}
\title{Return function to interpolate a atsu color palette}
\usage{
atsu_pal(palette, reverse, ...)
}
\arguments{
\item{palette}{Character name of palette in atsu_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a atsu color palette
}
