% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{atsu_palettes}
\alias{atsu_palettes}
\title{Create custom palettes using ATSU colors}
\format{
An object of class \code{list} of length 6.
}
\usage{
atsu_palettes
}
\description{
Create custom palettes using ATSU colors
}
\keyword{datasets}
