% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{A function to remove unwanted characters in a character vector}
\usage{
fix_names(x, patterns, replacements, case)
}
\arguments{
\item{x}{Character vector to be changed}

\item{patterns}{Character vector of patterns to find given as c("pattern")}

\item{replacements}{Character vector of replacements for patterns given as c("replacements")}

\item{case}{Character vector of either "lower" or "upper" for the output case}
}
\description{
A function to remove unwanted characters in a character vector
}
