% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{rand_pal}
\alias{rand_pal}
\title{Function to randomly select n colors from atsu_cols}
\usage{
rand_pal(n)
}
\arguments{
\item{n}{Numeric value of length of colors to sample}
}
\description{
Function to randomly select n colors from atsu_cols
}
