% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_atsu}
\alias{theme_atsu}
\alias{theme_atsu_clean}
\alias{labs_atsu}
\alias{clean_theme}
\alias{clean_table_theme}
\title{Publication ready theme from ggpubr}
\usage{
theme_atsu(
  base_size = 20,
  base_family = "",
  border = FALSE,
  margin = TRUE,
  legend = c("top", "bottom", "left", "right", "none"),
  x.text.angle = 0
)

theme_atsu_clean(base_size = 20, base_family = "", flip = FALSE)

labs_atsu(base_size = 20, base_family = "")

clean_theme()

clean_table_theme()
}
\arguments{
\item{base_size}{base font size given in pts}

\item{base_family}{base family}

\item{border}{logical if a border should be used}

\item{margin}{logical if margin should be used}

\item{legend}{location of legend placement}

\item{x.text.angle}{angle of x axis text}

\item{flip}{logical if x and y axis should be flipped
Basic theme for ATSU}
}
\description{
Publication ready theme from ggpubr

ATSU clean theme

Label theme

Add a clean theme

Add clean table theme
}
