% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_fill_atsu}
\alias{scale_fill_atsu}
\title{Fill scale constructor for atsu colors}
\usage{
scale_fill_atsu(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in atsu_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}

\item{random}{Boolean indicating whether to randomly choose colors from selected palette}
}
\description{
Fill scale constructor for atsu colors
}
